merge

+++ python/Security/Encrypt/encrypt.py
@@ -24,8 +24,9 @@ def encrypt(arg1, arg2):
         st = os.stat(arg2)
         fileSize = st.st_size
         fileSize = sys.getsizeof(plainText)
-        iteration = fileSize/16 
+        iteration = int(fileSize/16) 
         bytesLeft = fileSize%16
+
         keyVal = keyFile.read(16)
         for i in range(0, iteration-1):
 
@@ -33,6 +34,7 @@ def encrypt(arg1, arg2):
                 print ("Key Read Error")
 
             else:
+
                 start = i*16;
                 encryptedText = bytearray(16)
                 encryptedText = aesEncrypt(plainText[start:(start+16)], keyVal)
